publicise OgmaData and AsType
